# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Lathi API
  version: "1.0.0"
  description: API Documentation for Lathi - Javanese Visual Novel Learning Platform
servers:
  - url: "http://localhost:8080/api/v1"
    description: Local development server
  - url: "https://lathi.bccdev.id/api/v1"
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "success"
        data:
          oneOf:
            - type: object
            - type: "null"
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - type
            - message
            - status
          properties:
            type:
              type: string
            message:
              type: string
            detail:
              type: string
            status:
              type: integer
            fields:
              type: object
              additionalProperties:
                type: string

    # auth schemas
...
    StoryActionRequest:
      type: object
      required:
        - chapter_id
        - slide_id
      properties:
        chapter_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        slide_id:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440001"
        choice_index:
          oneOf:
            - type: integer
            - type: "null"
          example: 0
    HistoryEntry:
      type: object
      properties:
        speaker:
          type: string
          example: "Andi"
        text:
          type: string
          example: "Sugeng siang, Bu."
        is_user:
          type: boolean
          example: false
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    CharacterOnScreen:
      type: object
      properties:
        name:
          type: string
          example: "Andi"
        image_url:
          type: string
          example: "https://storage.lathi.id/chars/andi_happy.webp"
        is_active:
          type: boolean
          example: true
    ChapterListResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          example: "Warung Mie Andi"
        description:
          type: string
          example: "Andi ketemu Sekar neng warung, lan ngerti yen kudu ketemu Bapake."
        cover_image_url:
          type: string
          example: "https://storage.lathi.id/chapters/ch1_cover.webp"
        order_index:
          type: integer
          example: 1
        is_locked:
          type: boolean
          example: false
        is_completed:
          type: boolean
          example: false
    VocabItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        word_krama:
          type: string
          example: "sowan"
        word_ngoko:
          type: string
          example: "teka/mertamu"
        word_indo:
          type: string
          example: "berkunjung (hormat)"
    ChoiceItemResponse:
      type: object
      properties:
        index:
          type: integer
          example: 0
        text:
          type: string
          example: "Sugeng siang, Pak."
    SlideItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        background_image_url:
          type: string
          example: "https://storage.lathi.id/bg/warmindo.webp"
        characters:
          type: array
          items:
            $ref: "#/components/schemas/CharacterOnScreen"
        speaker_name:
          type: string
          example: "Narator"
        content:
          type: string
          example: "Wanci sonten ing kutha Surabaya..."
        next_slide_id:
          oneOf:
            - type: string
              format: uuid
            - type: "null"
          example: "660e8400-e29b-41d4-a716-446655440001"
        vocabularies:
          type: array
          items:
            $ref: "#/components/schemas/VocabItemResponse"
        choices:
          type: array
          items:
            $ref: "#/components/schemas/ChoiceItemResponse"
    ChapterContentResponse:
      type: object
      properties:
        chapter_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        slides:
          type: array
          items:
            $ref: "#/components/schemas/SlideItemResponse"
    UserSessionResponse:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        current_slide_id:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440001"
        current_hearts:
          type: integer
          example: 3
        is_game_over:
          type: boolean
          example: false
        is_completed:
          type: boolean
          example: false
        history_log:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"
    StoryActionResponse:
      type: object
      properties:
        is_game_over:
          type: boolean
          example: false
        is_completed:
          type: boolean
          example: false
        message:
          type: string
          example: ""
        remaining_hearts:
          type: integer
          example: 3
        next_slide_id:
          oneOf:
            - type: string
              format: uuid
            - type: "null"
          example: "660e8400-e29b-41d4-a716-446655440001"
        history_log:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"

    # dictionary schemas
...
  responses:
...
    ErrChaptersUnauthorized:
      description: Unauthorized - User not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              type: "unauthorized"
              message: "Kamu belum login, yuk login dulu"
              detail: "Kamu belum login, yuk login dulu"
              status: 401
    ErrChaptersInternal:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              type: "internal_error"
              message: "Coba lagi nanti ya!"
              detail: "Coba lagi nanti ya!"
              status: 500

    # /stories/chapters/:id/content errors
    ErrChapterContentBadRequest:
      description: Bad request - Invalid chapter ID parameter
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              type: "bad_request"
              message: "Data yang dikirimkan salah"
              detail: "Parameter 'id' harus berupa UUID yang valid"
              status: 400
    ErrChapterContentUnauthorized:
      description: Unauthorized - User not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              type: "unauthorized"
              message: "Kamu belum login, yuk login dulu"
              detail: "Kamu belum login, yuk login dulu"
              status: 401
    ErrChapterContentNotFound:
      description: Not found - Chapter not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              type: "not_found"
              message: "Data ga ditemukan"
              detail: "Chapter ini ga ketemu"
              status: 404
    ErrChapterContentInternal:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              type: "internal_error"
              message: "Coba lagi nanti ya!"
              detail: "Coba lagi nanti ya!"
              status: 500

    # /stories/chapters/:id/session errors
    ErrSessionBadRequest:
      description: Bad request - Invalid chapter ID parameter
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              type: "bad_request"
              message: "Data yang dikirimkan salah"
              detail: "Parameter 'id' harus berupa UUID yang valid"
              status: 400
    ErrSessionUnauthorized:
      description: Unauthorized - User not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              type: "unauthorized"
              message: "Kamu belum login, yuk login dulu"
              detail: "Kamu belum login, yuk login dulu"
              status: 401
    ErrSessionInternal:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              type: "internal_error"
              message: "Coba lagi nanti ya!"
              detail: "Coba lagi nanti ya!"
              status: 500

    # /stories/chapters/:id/start errors
    ErrStartSessionBadRequest:
      description: Bad request - Invalid chapter ID parameter
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              type: "bad_request"
              message: "Data yang dikirimkan salah"
              detail: "Parameter 'id' harus berupa UUID yang valid"
              status: 400
    ErrStartSessionUnauthorized:
      description: Unauthorized - User not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              type: "unauthorized"
              message: "Kamu belum login, yuk login dulu"
              detail: "Kamu belum login, yuk login dulu"
              status: 401
    ErrStartSessionNotFound:
      description: Not found - Chapter not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              type: "not_found"
              message: "Data ga ditemukan"
              detail: "Chapter ini ga ketemu"
              status: 404
    ErrStartSessionInternal:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            generalError:
              summary: General internal error
              value:
                success: false
                error:
                  type: "internal_error"
                  message: "Coba lagi nanti ya!"
                  detail: "Coba lagi nanti ya!"
                  status: 500
            noContent:
              summary: Chapter has no content/slides
              value:
                success: false
                error:
                  type: "internal_error"
                  message: "Coba lagi nanti ya!"
                  detail: "Chapter ini belum punya konten"
                  status: 500

    # /stories/action errors
    ErrActionBadRequest:
      description: Bad request - Multiple scenarios (parse error, validation errors, session errors)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            parseError:
              summary: Failed to parse request body
              value:
                success: false
                error:
                  type: "bad_request"
                  message: "Data yang dikirimkan salah"
                  detail: "Data yang kamu kirim belum pas, coba cek lagi ya"
                  status: 400
            sessionNotStarted:
              summary: Session not started yet
              value:
                success: false
                error:
                  type: "bad_request"
                  message: "Data yang dikirimkan salah"
                  detail: "Kamu belum memulai chapter ini, yuk mulai dulu ya!"
                  status: 400
            sessionEnded:
              summary: Session already ended (game over or completed)
              value:
                success: false
                error:
                  type: "bad_request"
                  message: "Data yang dikirimkan salah"
                  detail: "Permainan udah selesai, coba mulai lagi ya!"
                  status: 400
            choiceRequired:
              summary: Choice is required but not provided
              value:
                success: false
                error:
                  type: "bad_request"
                  message: "Data yang dikirimkan salah"
                  detail: "Kamu harus milih salah satu pilihan yang ada"
                  status: 400
            noChoiceAvailable:
              summary: Choice provided but slide has no choices
              value:
                success: false
                error:
                  type: "bad_request"
                  message: "Data yang dikirimkan salah"
                  detail: "Slide ini ga punya pilihan buat dipilih"
                  status: 400
            invalidChoice:
              summary: Invalid choice index
              value:
                success: false
                error:
                  type: "bad_request"
                  message: "Data yang dikirimkan salah"
                  detail: "Pilihanmu ga valid"
                  status: 400
    ErrActionValidation:
      description: Validation error - Invalid input fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              type: "validation_error"
              message: "Ups, ada data yang ga sesuai nih"
              status: 422
              fields:
                chapter_id: "required"
                slide_id: "uuid"
    ErrActionUnauthorized:
      description: Unauthorized - User not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              type: "unauthorized"
              message: "Kamu belum login, yuk login dulu"
              detail: "Kamu belum login, yuk login dulu"
              status: 401
    ErrActionNotFound:
      description: Not found - Slide not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              type: "not_found"
              message: "Data ga ditemukan"
              detail: "Slide ga ketemu"
              status: 404
    ErrActionInternal:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              type: "internal_error"
              message: "Coba lagi nanti ya!"
              detail: "Coba lagi nanti ya!"
              status: 500

    # /dictionaries errors
...
paths:
...
  /stories/chapters:
    get:
      tags:
        - Story
      summary: Get Chapter List
      description: Get list of all available chapters with user progress (locked/completed status).
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK - Chapter list retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      message:
                        example: "Daftar chapter berhasil dimuat"
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ChapterListResponse"
              example:
                success: true
                message: "Daftar chapter berhasil dimuat"
                data:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    title: "Ana Kabar Kaget!"
                    description: "Andi kudu nekat budhal nang Tulungagung demi mjuangne tresnane."
                    cover_image_url: "https://storage.lathi.id/chapters/ch1_cover.webp"
                    order_index: 1
                    is_locked: false
                    is_completed: true
                  - id: "550e8400-e29b-41d4-a716-446655440001"
                    title: "Sinau Dadi Priyayi"
                    description: "Andi meguru tata krama nang Pakdhe Joyo ben ora ngisin-ngisini."
                    cover_image_url: "https://storage.lathi.id/chapters/ch2_cover.webp"
                    order_index: 2
                    is_locked: true
                    is_completed: false
        "401":
          $ref: "#/components/responses/ErrChaptersUnauthorized"
        "500":
          $ref: "#/components/responses/ErrChaptersInternal"
  /stories/chapters/{id}/content:
    get:
      tags:
        - Story
      summary: Get Chapter Content
      description: Get complete content of a chapter including all slides, characters, vocabularies, and choices.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Chapter UUID
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "200":
          description: OK - Chapter content retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      message:
                        example: "Konten chapter berhasil dimuat"
                      data:
                        $ref: "#/components/schemas/ChapterContentResponse"
              example:
                success: true
                message: "Konten chapter berhasil dimuat"
                data:
                  chapter_id: "550e8400-e29b-41d4-a716-446655440000"
                  slides:
                    - id: "660e8400-e29b-41d4-a716-446655440001"
                      background_image_url: "https://storage.lathi.id/bg/warmindo.webp"
                      characters:
                        - name: "Andi"
                          image_url: "https://storage.lathi.id/chars/andi_neutral.webp"
                          is_active: true
                        - name: "Sekar"
                          image_url: "https://storage.lathi.id/chars/sekar_happy.webp"
                          is_active: false
                      speaker_name: "Narator"
                      content: "Wanci sonten ing kutha Surabaya..."
                      next_slide_id: "660e8400-e29b-41d4-a716-446655440002"
                      vocabularies:
                        - id: "770e8400-e29b-41d4-a716-446655440000"
                          word_krama: "sowan"
                          word_ngoko: "teka/mertamu"
                          word_indo: "berkunjung (hormat)"
                      choices:
                        - index: 0
                          text: "Sugeng siang, Pak."
                        - index: 1
                          text: "Assalamualaikum, Pak."
        "400":
          $ref: "#/components/responses/ErrChapterContentBadRequest"
        "401":
          $ref: "#/components/responses/ErrChapterContentUnauthorized"
        "404":
          $ref: "#/components/responses/ErrChapterContentNotFound"
        "500":
          $ref: "#/components/responses/ErrChapterContentInternal"
  /stories/chapters/{id}/session:
    get:
      tags:
        - Story
      summary: Get User Session
      description: Retrieve user's current progress/session for a specific chapter. Returns null if user hasn't started the chapter yet.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Chapter UUID
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "200":
          description: OK - Session retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      message:
                        example: "Progressmu berhasil dipulihkan"
                      data:
                        oneOf:
                          - $ref: "#/components/schemas/UserSessionResponse"
                          - type: "null"
              examples:
                withSession:
                  summary: User has started the chapter
                  value:
                    success: true
                    message: "Progressmu berhasil dipulihkan"
                    data:
                      session_id: "550e8400-e29b-41d4-a716-446655440000"
                      current_slide_id: "660e8400-e29b-41d4-a716-446655440001"
                      current_hearts: 2
                      is_game_over: false
                      is_completed: false
                      history_log:
                        - speaker: "Narator"
                          text: "Wanci sonten ing kutha Surabaya..."
                          is_user: false
                          timestamp: "2024-01-15T10:30:00Z"
                        - speaker: "Andi"
                          text: "Sugeng siang, Pak."
                          is_user: true
                          timestamp: "2024-01-15T10:30:15Z"
                noSession:
                  summary: User hasn't started the chapter
                  value:
                    success: true
                    message: "Progressmu berhasil dipulihkan"
                    data: null
        "400":
          $ref: "#/components/responses/ErrSessionBadRequest"
        "401":
          $ref: "#/components/responses/ErrSessionUnauthorized"
        "500":
          $ref: "#/components/responses/ErrSessionInternal"
  /stories/chapters/{id}/start:
    post:
      tags:
        - Story
      summary: Start New Session
      description: Start a new gameplay session for a specific chapter. Initializes progress with 3 hearts and sets current slide to the first slide.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Chapter UUID
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "200":
          description: OK - Session started successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      message:
                        example: "Permainan dimulai, semangat ya!"
                      data:
                        type: "null"
                        example: null
        "400":
          $ref: "#/components/responses/ErrStartSessionBadRequest"
        "401":
          $ref: "#/components/responses/ErrStartSessionUnauthorized"
        "404":
          $ref: "#/components/responses/ErrStartSessionNotFound"
        "500":
          $ref: "#/components/responses/ErrStartSessionInternal"
  /stories/action:
    post:
      tags:
        - Story
      summary: Submit Story Action
      description: |
        Submit user action (next slide or choice selection) during gameplay.
        - Updates session state (hearts, history log, current slide)
        - Handles choice selection with mood impact
        - Detects game over (hearts <= 0) or completion (no next slide)
        - Unlocks vocabularies when encountered
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoryActionRequest"
            examples:
              nextSlide:
                summary: Move to next slide (no choice)
                value:
                  chapter_id: "550e8400-e29b-41d4-a716-446655440000"
                  slide_id: "660e8400-e29b-41d4-a716-446655440001"
                  choice_index: null
              selectChoice:
                summary: Select a choice option
                value:
                  chapter_id: "550e8400-e29b-41d4-a716-446655440000"
                  slide_id: "660e8400-e29b-41d4-a716-446655440001"
                  choice_index: 0
      responses:
        "200":
          description: OK - Action processed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      message:
                        example: "Aksimu berhasil diproses!"
                      data:
                        $ref: "#/components/schemas/StoryActionResponse"
              examples:
                normalProgress:
                  summary: Normal progress (hearts remaining)
                  value:
                    success: true
                    message: "Aksimu berhasil diproses!"
                    data:
                      is_game_over: false
                      is_completed: false
                      message: ""
                      remaining_hearts: 3
                      next_slide_id: "660e8400-e29b-41d4-a716-446655440002"
                      history_log:
                        - speaker: "Narator"
                          text: "Wanci sonten ing kutha Surabaya..."
                          is_user: false
                          timestamp: "2024-01-15T10:30:00Z"
                        - speaker: "Andi"
                          text: "Sugeng siang, Pak."
                          is_user: true
                          timestamp: "2024-01-15T10:30:15Z"
                gameOver:
                  summary: Game over (hearts depleted)
                  value:
                    success: true
                    message: "Aksimu berhasil diproses!"
                    data:
                      is_game_over: true
                      is_completed: false
                      message: "Pak Broto kuciwo karo omonganmu. Coba maneh ya!"
                      remaining_hearts: 0
                      next_slide_id: "660e8400-e29b-41d4-a716-446655440002"
                      history_log:
                        - speaker: "Andi"
                          text: "Mas Andi aja ngono..."
                          is_user: true
                          timestamp: "2024-01-15T10:35:00Z"
                completed:
                  summary: Chapter completed
                  value:
                    success: true
                    message: "Aksimu berhasil diproses!"
                    data:
                      is_game_over: false
                      is_completed: true
                      message: "Sugeng! Sampeyan wis rampung crita iki."
                      remaining_hearts: 2
                      next_slide_id: null
                      history_log:
                        - speaker: "Narator"
                          text: "TAMAT."
                          is_user: false
                          timestamp: "2024-01-15T10:40:00Z"
        "400":
          $ref: "#/components/responses/ErrActionBadRequest"
        "422":
          $ref: "#/components/responses/ErrActionValidation"
        "401":
          $ref: "#/components/responses/ErrActionUnauthorized"
        "404":
          $ref: "#/components/responses/ErrActionNotFound"
        "500":
          $ref: "#/components/responses/ErrActionInternal"

  # dictionary endpoints
...